requirements.txt
yfinance
pandas
numpy
scikit-learn
fastapi
uvicorn[standard]
pydantic
joblib
pyyaml
pyarrow
matplotlib
